Added improvements to HalDashboard
Added a key mechanism where different modules can request and reserve keys to different lines on the dashboard. Text can only be written using a key. This solves the problem where multiple modules are writing to the same lines. The previous method of free-for-all display is still supported, where a key is temporarily reserved and subsequently released. This makes this improvement backwards compatible with the existing implementation of displaying text.