Bug fixes, logging, and optimizations.
Hopefully fixed the bug where the code won't work without a redeploy.

Added logging to show the unscaled and scaled motion profiles. There's some problem where the talon is reporting points that are different from the targets. Hopefully this should highlight it.

Switched from array of statuses to 2 statuses, so it's hopefully clearer.